feat: Rename MainViewModel to ThemeViewModel
