+++ app/src/main/java/com/kongjak/koreatechboard/ui/activity/SearchActivity.kt
@@ -3,16 +3,23 @@ package com.kongjak.koreatechboard.ui.activity
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.activity.viewModels
+import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.Modifier
 import com.kongjak.koreatechboard.ui.search.SearchScreen
 import com.kongjak.koreatechboard.ui.theme.KoreatechBoardTheme
+import com.kongjak.koreatechboard.ui.viewmodel.ThemeViewModel
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class SearchActivity : ComponentActivity() {
+
+    private val themeViewModel: ThemeViewModel by viewModels()
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
@@ -21,7 +28,13 @@ class SearchActivity : ComponentActivity() {
         val title = intent.getStringExtra("title")!!
 
         setContent {
-            KoreatechBoardTheme {
+            val isDynamicColor by themeViewModel.isDynamicTheme.observeAsState(true)
+            val isDarkTheme by themeViewModel.isDarkTheme.observeAsState()
+            KoreatechBoardTheme(
+                dynamicColor = isDynamicColor,
+                darkTheme = isDarkTheme ?: isSystemInDarkTheme()
+            )
+            {
                 // A surface container using the 'background' color from the theme
                 Surface(
                     modifier = Modifier.fillMaxSize(),
